// src/app/pages/vehiculoreparandos/components/vehiculoreparandos-table/vehiculoreparandos.interface.ts
export interface VehiculoreparandosInterface {

    idvehiculoreparando?: number;
    taller_idtaller: number;
    enviotaller_idenviotaller: number;
    fechaIngresa: string;
    mecanico_idmecanico: number;
    fechaSalida: string;
    fechaEstimada: string;
    inventario:string;
    motivo:string;
    status: string;
    orden:string;
    permisotaxiasignado_idpermisotaxiasignado:number;
}